Add files via upload
Trying all the different uses of Python. I used an very popular module called Pillow (formerly know as PIL).
It was all fun until I stumbled in an bug that was causing files not to actually change extension but just change the add a .png before .jpg (or any file extension) but like hour later i figured out over all it was pretty great now if i have an image to crop or change jpeg or any other files types to png I can just lead that image to the that directory and just get it done with dipping my head in photoshop ui and function hell.