Add files via upload
little brief  about the code: I made an very basic python number guessing game it was not so special alot of tutorial is available on yt about the the uncommon thing in my project is level and difficulty and I also learnt how to use os module in python I primarily used os.system to clear the terminal throughout the game.
developing the industry quality skills little by little. I'm trying to be consistent and made projects on daily basis.